requirements.txt
flask==3.0.3
werkzeug==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
Pillow==10.3.0
itsdangerous==2.2.0
gunicorn==21.2.0
